backend/requirements.txt
flask
numpy
tensorflow
pillow
transformers
torch
pyserial
flask-cors 